Add ability to use configs to organize multi libraries easier
Before these changes, moc generator required include paths to be specified in other ways for the compiler. In this way, we can give the include paths by passing in include folders by inheritance.

The key in this commit is allowing users to use inherited build configs -- it isn't simply self-contained but setting up a config will save you a lot of time configuring other qt-related builds.